Refactor reports to separate concerns
